some refactoring, and DEX improvement started in order to add new NODES easily
